backend/requirements.txt
flask
python-dotenv
psycopg2-binary
sqlalchemy
flask-sqlalchemy